:construction: fin journée lundi

+++ src/main/java/fr/univartois/sae/raytracing/triplet/Point.java
@@ -22,7 +22,7 @@ public class Point {
      * @param x,y,z is a triplet
      */
 
-    public Point(int x,int y,int z){
+    public Point(double x,double y,double z){
         triplet=new Triplet(x,y,z);
     }
     /**
